pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-cloud-vision>=3.9.0",
    "gunicorn>=23.0.0",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "pytesseract>=0.3.13",
    "requests>=2.32.3",
]
